rolled back for now methods moved from environment.py to reduce merge conflicts
