+++ common/survey_utilities.py
@@ -5,11 +5,12 @@ from random import randint
 
 from structlog import wrap_logger
 
+from acceptance_tests.features import environment
 from acceptance_tests.features.data_setup import display_unused_enrolment_code, generate_new_enrolment_code, \
     survey_enrolment, add_a_survey, social_change_case_status, social_disable_uac, social_find_case_by_postcode, \
     social_view_case_details
-from common import collection_exercise_utilities, respondent_utilities
 from common import common_utilities
+from common import respondent_utilities
 from config import Config
 from controllers import collection_exercise_controller, survey_controller
 
@@ -37,8 +38,8 @@ logger = wrap_logger(getLogger(__name__))
 # Non-standalone methods
 
 def setup_non_standalone_data_for_test():
-    collection_exercise_utilities.execute_collection_exercises()
-    collection_exercise_utilities.register_respondent(survey_id='cb8accda-6118-4d3b-85a3-149e28960c54', period='201801',
+    environment.execute_collection_exercises()
+    environment.register_respondent(survey_id='cb8accda-6118-4d3b-85a3-149e28960c54', period='201801',
                                                       username=Config.RESPONDENT_USERNAME, ru_ref=49900000001)
 
 
@@ -143,8 +144,8 @@ def create_social_collection_exercise_for_test(context, survey_id, period, ru_re
 
     logger.info('Creating Social Collection Exercise', survey_id=survey_id, period=period)
 
-    user_description = collection_exercise_utilities.make_user_description(ce_name, is_social_survey(survey_type), 50)
-    dates = collection_exercise_utilities.generate_social_collection_exercise_dates()
+    user_description = environment.make_user_description(ce_name, is_social_survey(survey_type), 50)
+    dates = environment.generate_social_collection_exercise_dates()
 
     iac = collection_exercise_controller.create_and_execute_social_collection_exercise(context, survey_id, period,
                                                                                        user_description, dates,
@@ -164,8 +165,8 @@ def create_business_collection_exercise_for_test(survey_id, period, ru_ref, ce_n
 
     logger.info('Creating Business Collection Exercise', survey_id=survey_id, period=period)
 
-    user_description = collection_exercise_utilities.make_user_description(ce_name, is_social_survey(survey_type), 50)
-    dates = collection_exercise_utilities.generate_collection_exercise_dates_from_period(period)
+    user_description = environment.make_user_description(ce_name, is_social_survey(survey_type), 50)
+    dates = environment.generate_collection_exercise_dates_from_period(period)
 
     iac = collection_exercise_controller.create_and_execute_collection_exercise_with_unique_sample(survey_id, period,
                                                                                                    user_description,
@@ -187,9 +188,9 @@ def create_respondent_enrolled_in_the_test_survey(context):
     respondent_utilities.create_respondent(user_name=user_name, enrolment_code=context.iac)['id']
     respondent_utilities.create_respondent_user_login_account(user_name)
 
-    case = collection_exercise_utilities.find_case_by_enrolment_code(context.iac)
+    case = environment.find_case_by_enrolment_code(context.iac)
 
-    context.iac = collection_exercise_utilities.generate_new_enrolment_code(case['id'], case['partyId'])
+    context.iac = environment.generate_new_enrolment_code(case['id'], case['partyId'])
 
 
 def create_respondent_not_enrolled_in_the_test_survey(context, wait_for_live_collection_exercise=False):
